+++ packages/retail-ui/components/ComboBox/__tests__/ComboBox-test.tsx
@@ -1,5 +1,9 @@
 // tslint:disable:jsx-no-lambda
 import * as React from 'react';
+import { CustomComboBoxLocaleHelper } from '../../CustomComboBox/locale';
+import { LangCodes } from '../../LocaleProvider';
+import { defaultLangCode } from '../../LocaleProvider/constants';
+import LocaleProvider from '../../LocaleProvider/LocaleProvider';
 import ComboBox, { ComboBoxProps } from '../ComboBox';
 import { mount, ReactWrapper } from 'enzyme';
 import InputLikeText from '../../internal/InputLikeText';
@@ -8,7 +12,7 @@ import Menu from '../../Menu/Menu';
 import { delay } from 'retail-ui/lib/utils';
 import CustomComboBox, { DELAY_BEFORE_SHOW_LOADER, LOADER_SHOW_TIME } from '../../CustomComboBox/CustomComboBox';
 import ComboBoxView from '../../CustomComboBox/ComboBoxView';
-import { Effect } from '../../CustomComboBox/reducer/default';
+import { Effect } from '../../CustomComboBox/CustomComboBoxReducer';
 import { ComboBoxRequestStatus } from '../../CustomComboBox/CustomComboBoxTypes';
 
 function clickOutside() {
@@ -37,11 +41,11 @@ describe('ComboBox', () => {
   });
 
   it('renders', () => {
-    mount<ComboBox<any>>(<ComboBox />);
+    mount<ComboBox<any>>(<ComboBox getItems={() => Promise.resolve([])} />);
   });
 
   it('focuses on focus call', () => {
-    const wrapper = mount<ComboBox<any>>(<ComboBox />);
+    const wrapper = mount<ComboBox<any>>(<ComboBox getItems={() => Promise.resolve([])} />);
     wrapper.instance().focus();
     expect(wrapper.getDOMNode().contains(document.activeElement)).toBeTruthy();
   });
@@ -189,7 +193,9 @@ describe('ComboBox', () => {
   it('calls onChange if onUnexpectedInput return non-nullary value', async () => {
     const values = [null, undefined, 'one'];
     const onChange = jest.fn();
-    const wrapper = mount<ComboBox<string>>(<ComboBox onChange={onChange} onUnexpectedInput={value => value} />);
+    const wrapper = mount<ComboBox<string>>(
+      <ComboBox onChange={onChange} onUnexpectedInput={value => value} getItems={() => Promise.resolve([])} />,
+    );
 
     while (values.length) {
       wrapper.instance().focus();
@@ -205,7 +211,7 @@ describe('ComboBox', () => {
 
   it('calls onFocus on focus', async () => {
     const onFocus = jest.fn();
-    const wrapper = mount<ComboBox<any>>(<ComboBox onFocus={onFocus} />);
+    const wrapper = mount<ComboBox<any>>(<ComboBox onFocus={onFocus} getItems={() => Promise.resolve([])} />);
 
     wrapper.find('[tabIndex=0]').simulate('focus');
 
@@ -330,7 +336,7 @@ describe('ComboBox', () => {
   });
 
   it("don't focus on error and value change", () => {
-    const wrapper = mount<ComboBox<any>>(<ComboBox />);
+    const wrapper = mount<ComboBox<any>>(<ComboBox getItems={() => Promise.resolve([])} />);
 
     wrapper.setProps({ value: { label: '1' }, error: true });
     wrapper.update();
@@ -339,7 +345,9 @@ describe('ComboBox', () => {
   });
 
   it('clear value if onUnexpectedInput return null', async () => {
-    const wrapper = mount<ComboBox<any>>(<ComboBox onUnexpectedInput={() => null} />);
+    const wrapper = mount<ComboBox<any>>(
+      <ComboBox onUnexpectedInput={() => null} getItems={() => Promise.resolve([])} />,
+    );
 
     wrapper.instance().focus();
     wrapper.update();
@@ -462,11 +470,20 @@ describe('ComboBox', () => {
     };
 
     it('in default mode', () => {
-      check(mount<ComboBox<any>>(<ComboBox value={VALUES[0]} />));
+      check(mount<ComboBox<any>>(<ComboBox value={VALUES[0]} getItems={() => Promise.resolve(VALUES)} />));
     });
 
     it('in autocomplete mode', () => {
-      check(mount<ComboBox<any>>(<ComboBox value={VALUES[0]} autocomplete={true} />));
+      check(
+        mount<ComboBox<any>>(
+          <ComboBox
+            value={VALUES[0]}
+            drawArrow={false}
+            searchOnFocus={false}
+            getItems={() => Promise.resolve(VALUES)}
+          />,
+        ),
+      );
     });
   });
 
@@ -486,18 +503,24 @@ describe('ComboBox', () => {
     };
 
     it('in default mode', async () => {
-      check(mount<ComboBox<any>>(<ComboBox value={value} />));
+      check(mount<ComboBox<any>>(<ComboBox value={value} getItems={() => Promise.resolve([value])} />));
     });
 
     it('in autocomplete mode', async () => {
-      check(mount<ComboBox<any>>(<ComboBox value={value} autocomplete={true} />));
+      check(
+        mount<ComboBox<any>>(
+          <ComboBox value={value} drawArrow={false} searchOnFocus={false} getItems={() => Promise.resolve([value])} />,
+        ),
+      );
     });
   });
 
   it('does not do search on focus in autocomplete mode', async () => {
-    const VALUE = { value: 1, label: 'one' };
+    const value = { value: 1, label: 'one' };
     const getItems = jest.fn();
-    const wrapper = mount<ComboBox<any>>(<ComboBox getItems={getItems} value={VALUE} autocomplete={true} />);
+    const wrapper = mount<ComboBox<any>>(
+      <ComboBox getItems={getItems} value={value} drawArrow={false} searchOnFocus={false} />,
+    );
 
     wrapper.instance().focus();
     await delay(0);
@@ -508,7 +531,7 @@ describe('ComboBox', () => {
   });
 
   it('reset', () => {
-    const wrapper = mount<ComboBox<any>>(<ComboBox />);
+    const wrapper = mount<ComboBox<any>>(<ComboBox getItems={() => Promise.resolve([])} />);
 
     wrapper.instance().focus();
     wrapper.update();
@@ -567,10 +590,10 @@ describe('ComboBox', () => {
   });
 
   describe('open/close methods', () => {
-    let wrapper: ReactWrapper<{}, {}, ComboBox<any>>;
+    let wrapper: ReactWrapper<ComboBoxProps<any>, {}, ComboBox<any>>;
 
     beforeEach(() => {
-      wrapper = mount<ComboBox<any>>(<ComboBox />);
+      wrapper = mount<ComboBox<any>>(<ComboBox getItems={() => Promise.resolve([])} />);
       wrapper.instance().open();
       wrapper.update();
     });
@@ -596,7 +619,7 @@ describe('ComboBox', () => {
     const VALUE = { value: 1, label: 'one' };
     let getItems: jest.Mock<Promise<string[]>>;
     let promise: Promise<{}>;
-    let wrapper: ReactWrapper<{}, {}, ComboBox<typeof VALUE>>;
+    let wrapper: ReactWrapper<ComboBoxProps<typeof VALUE>, {}, ComboBox<typeof VALUE>>;
 
     beforeEach(() => {
       [getItems, promise] = searchFactory(Promise.resolve(['one']));
@@ -658,7 +681,7 @@ describe('ComboBox', () => {
 
   describe('click on input', () => {
     const VALUE = { value: 1, label: 'one' };
-    type TComboBoxWrapper = ReactWrapper<{}, {}, ComboBox<typeof VALUE>>;
+    type TComboBoxWrapper = ReactWrapper<ComboBoxProps<typeof VALUE>, {}, ComboBox<typeof VALUE>>;
     const clickOnInput = (comboboxWrapper: TComboBoxWrapper) => {
       comboboxWrapper.update();
       comboboxWrapper.find('input').simulate('click');
@@ -698,7 +721,9 @@ describe('ComboBox', () => {
 
     describe('in autocomplete mode', () => {
       beforeEach(() => {
-        wrapper = mount<ComboBox<typeof VALUE>>(<ComboBox autocomplete={true} getItems={getItems} value={VALUE} />);
+        wrapper = mount<ComboBox<typeof VALUE>>(
+          <ComboBox drawArrow={false} searchOnFocus={false} getItems={getItems} value={VALUE} />,
+        );
         wrapper.instance().focus();
         getItems.mockClear();
       });
@@ -1076,6 +1101,86 @@ describe('ComboBox', () => {
         loading: true,
         opened: true,
       });
+    });
+  });
+
+  describe('Locale', () => {
+    let search: jest.Mock<Promise<any>>;
+    let promise: Promise<any>;
+    let wrapper: ReactWrapper;
+    beforeEach(() => {
+      [search, promise] = searchFactory(Promise.resolve(null));
+    });
+    const focus = async (): Promise<void> => {
+      (wrapper.find(ComboBox).instance() as ComboBox<any>).focus();
+      await promise;
+      wrapper.update();
+    };
+
+    it('render without LocaleProvider', async () => {
+      wrapper = mount(<ComboBox getItems={search} />);
+      const expectedText = CustomComboBoxLocaleHelper.get(defaultLangCode).notFound;
+
+      await focus();
+
+      expect(wrapper.find(MenuItem).text()).toBe(expectedText);
+    });
+
+    it('render default locale', async () => {
+      wrapper = mount(
+        <LocaleProvider>
+          <ComboBox getItems={search} />
+        </LocaleProvider>,
+      );
+      const expectedText = CustomComboBoxLocaleHelper.get(defaultLangCode).notFound;
+
+      await focus();
+
+      expect(wrapper.find(MenuItem).text()).toBe(expectedText);
+    });
+
+    it('render correct locale when set langCode', async () => {
+      wrapper = mount(
+        <LocaleProvider langCode={LangCodes.en_EN}>
+          <ComboBox getItems={search} />
+        </LocaleProvider>,
+      );
+      const expectedText = CustomComboBoxLocaleHelper.get(LangCodes.en_EN).notFound;
+
+      await focus();
+
+      expect(wrapper.find(MenuItem).text()).toBe(expectedText);
+    });
+
+    it('render custom locale', async () => {
+      const customText = 'custom notFound';
+      wrapper = mount(
+        <LocaleProvider
+          locale={{
+            ComboBox: { notFound: customText },
+          }}
+        >
+          <ComboBox getItems={search} />
+        </LocaleProvider>,
+      );
+
+      await focus();
+
+      expect(wrapper.find(MenuItem).text()).toBe(customText);
+    });
+
+    it('updates when langCode changes', async () => {
+      wrapper = mount(
+        <LocaleProvider>
+          <ComboBox getItems={search} />
+        </LocaleProvider>,
+      );
+      const expected = CustomComboBoxLocaleHelper.get(LangCodes.en_EN).notFound;
+
+      wrapper.setProps({ langCode: LangCodes.en_EN });
+      await focus();
+
+      expect(wrapper.find(MenuItem).text()).toBe(expected);
     });
   });
 });