+++ packages/retail-ui/components/Spinner/Spinner.tsx
@@ -2,10 +2,12 @@ import classnames from 'classnames';
 import * as React from 'react';
 
 import * as PropTypes from 'prop-types';
+import { locale } from '../LocaleProvider/decorators';
+import { SpinnerLocale, SpinnerLocaleHelper } from './locale';
+import { sizeMaps, svgAnimateSupport, types } from './settings';
 
 import styles from './Spinner.less';
 import SpinnerFallback from './SpinnerFallback';
-import { types, sizeMaps, svgAnimateSupport } from './settings';
 
 export const SpinnerConfig = {
   hasSvgAnimationSupport: svgAnimateSupport(),
@@ -22,6 +24,8 @@ export interface SpinnerProps {
 /**
  * DRAFT - инлайн-лоадер
  */
+
+@locale('Spinner', SpinnerLocaleHelper)
 class Spinner extends React.Component<SpinnerProps> {
   public static propTypes = {
     /**
@@ -45,13 +49,14 @@ class Spinner extends React.Component<SpinnerProps> {
 
   public static defaultProps = {
     type: types.normal,
-    caption: 'Загрузка',
   };
 
   public static Types: typeof types;
 
+  private readonly locale!: SpinnerLocale;
+
   public render() {
-    const { type, caption, dimmed } = this.props;
+    const { type, caption = this.locale.loading, dimmed } = this.props;
     const verifiedType = sizeMaps[type!] ? type! : Spinner.defaultProps.type;
 
     return (