+++ packages/retail-ui/components/Toggle/Toggle.tsx
@@ -3,12 +3,9 @@ import * as PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Upgrades from '../../lib/Upgrades';
 
-import DefaultStyles = require('./Toggle.less');
-import FlatStyles = require('./Toggle.flat.less');
-
 const isFlatDesign = Upgrades.isFlatDesignEnabled();
 
-const styles = isFlatDesign ? FlatStyles : DefaultStyles;
+const styles = isFlatDesign ? require('./Toggle.flat.less') : require('./Toggle.less');
 
 let isListening: boolean;
 let tabPressed: boolean;