+++ packages/retail-ui/components/TopBar/locale/types.ts
@@ -1,0 +1,7 @@
+export interface TopBarLocale {
+  logout: string;
+  cabinetTitle: string;
+  cabinetSettings: string;
+  cabinetCertificates: string;
+  cabinetServices: string;
+}